+++ pages/about.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Layout from '@/components/common/Layout'
-import {Info} from '@/components/organisms/Alert'
+import {Info} from '@/components/molecules/Alert'
 
 const About: React.FC = () => {
   return (
